+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug Fixes
 
+- Qute - incorrect behavior for nested template records. See [#987](https://github.com/redhat-developer/quarkus-ls/issues/987).
 - Version 18 release duplicate key issue. See [#975](https://github.com/redhat-developer/quarkus-ls/issues/988).
 
 ## [0.19.0](https://github.com/redhat-developer/quarkus-ls/milestone/21?closed=1) (August 26, 2024)